feat: add basic i18n with DE routes
